Add basic authentication for admin actions
Worksighted would like authentication for admin actions such as managing users.

We have the authentication conditional to if there is a password set. We don't
want to have to log in if there is not a password. There is not a password for
our local configuration, so we don't want the authentication in that case.

[[Finishes #149859218]](https://www.pivotaltracker.com/story/show/149859218)